diff pick Signed-off-by: s30029175 <shiqiwei4@huawei.com>

+++ surface/src/buffer_queue.cpp
@@ -1068,7 +1068,7 @@ void BufferQueue::ReleaseDropBuffers(std::vector<BufferAndFence> &dropBuffers)
 
 void BufferQueue::DropBuffersByLevel(std::vector<BufferAndFence> &dropBuffers)
 {
-    if (dropFrameLevel_ <= 0 || dirtyList_.size() <= static_cast<size_t>(dropFrameLevel_)) {
+    if (dropFrameLevel_ <= 0 || dirtyList_.size() <= dropFrameLevel_) {
         return;
     }
